Function sentry_contrib_native::set_hook

pub fn set_hook(hook: Option<Box<dyn Fn(&PanicInfo) + Sync + Send + 'static>>)

Panic handler to send an event with the current stacktrace to Sentry.

This will not work properly if used with panic = "abort" because Shutdown is never unwound. To fix this make sure you make the panic handler itself call shutdown.

Examples

fn main() -> Result<()> {
    // pass original panic handler provided by rust to retain it's functionality
    set_hook(Some(std::panic::take_hook()));
    // it can also be removed
    set_hook(None);

    let _shutdown = Options::new().init()?;

    panic!("application panicked")
}

If you are using panic = "abort" make sure to call shutdown inside the panic handler.

set_hook(Some(Box::new(|_| shutdown())));