Module sentry_contrib_native::options

Sentry options implementation.

Structs

Options
The Sentry client options.

Shutdown
Automatically shuts down the Sentry client on drop.

Enums

Ownership
Represents the ownership status of Options.

Statics

GLOBAL_LOCK
Global lock for the following purposes:

Functions

global_read
Convenience function to get a read lock on GLOBAL_LOCK.

global_write
Convenience function to get a write lock on GLOBAL_LOCK.