Trait sentry_contrib_native::options::BeforeSend

pub trait BeforeSend: 'static + Send + Sync {
    fn before_send(&mut self, value: Value) -> Value;
}

Trait to help pass data to Options::set_before_send.

Examples

struct Filter {
    filtered: usize,
};

impl BeforeSend for Filter {
    fn before_send(&mut self, value: Value) -> Value {
        self.filtered += 1;
        // do something with the value and then return it
        value
    }
}

let mut options = Options::new();
options.set_before_send(Filter { filtered: 0 });
let _shutdown = options.init()?;

Required methods

fn before_send(&mut self, value: Value) -> Value

Before send callback.

Examples

struct Filter {
    filtered: usize,
};

impl BeforeSend for Filter {
    fn before_send(&mut self, value: Value) -> Value {
        self.filtered += 1;
        // do something with the value and then return it
        value
    }
}

Loading content...