Struct sentry_contrib_native::RawEnvelope

pub struct RawEnvelope(_);

Wrapper for the raw envelope that we should send to Sentry.

Examples

struct CustomTransport {
    #[cfg(feature = "transport-custom")]
    dsn: Dsn,
};

impl Transport for CustomTransport {
    fn send(&self, raw_envelope: RawEnvelope) {
        // get the `Event` that is being sent
        let event: Value = raw_envelope.event();
        // serialize it, maybe move this to another thread to prevent blocking
        let envelope: Envelope = raw_envelope.serialize();
        // or convert it into a `Request` right away!
        #[cfg(feature = "transport-custom")]
        let request: Request = raw_envelope.to_request(self.dsn.clone());
    }
}

Implementations

impl RawEnvelope

#[must_use = "`RawEnvelope::serialize` only converts it to an `Envelope`, this doesn't do anything until it is sent"]
pub fn serialize(&self) -> Envelope

Serialize a RawEnvelope into an Envelope.

#[must_use]
pub fn event(&self) -> Value

Yields the event that is being sent in the form of a Value.

#[must_use = "`Request` doesn't do anything until it is sent"]
pub fn to_request(&self, dsn: Dsn) -> Request

This is supported on crate feature transport-custom only.

Constructs a HTTP request for the provided RawEnvelope with a Dsn.

For more information see Envelope::into_request.