Module sentry_contrib_native::transport

Sentry custom transport implementation.

This can be used to send data to an upstream Sentry service in lieue of the built-in transports provided by the sentry-native library itself.

Structs

Dsn	feature="custom-transport"
Contains the pieces we need to send requests based on the DSN the user set on Options

Envelope
The actual body which transports send to Sentry.

Parts	feature="custom-transport"
Parts aquired from Dsn::into_parts.

RawEnvelope
Wrapper for the raw Envelope that we should send to Sentry.

Enums

Error	feature="custom-transport"
Sentry errors.

Shutdown
The return from Transport::shutdown, which determines if we tell the Sentry SDK if we were able to send all requests to the remote service or not in the allotted time.

State
Internal state of the Transport.

Constants

API_VERSION
Version of the Sentry API we can communicate with, AFAICT this is just hardcoded into sentry-native, so...two can play at that game!

ENVELOPE_MIME
The MIME type for Sentry envelopes.

SDK_USER_AGENT
SDK Version

Traits

Transport
Trait used to define a custom transport that Sentry can use to send events to a Sentry service.

Functions

send
Function to pass to sys::transport_new, which in turn calls the user defined one.

shutdown
Function to pass to sys::transport_set_shutdown_func, which in turn calls the user defined one.

startup
Function to pass to sys::transport_set_startup_func, which in turn calls the user defined one.

Type Definitions

Request	feature="custom-transport"
The http::Request request your Transport is expected to send.