Enum sentry_contrib_native::value::Value

pub enum Value {
    Null,
    Bool(bool),
    Int(i32),
    Double(f64),
    String(String),
    List(Vec<Value>),
    Map(BTreeMap<String, Value>),
}

Represents a Sentry protocol value.

Variants

Null

Null value.

Bool(bool)

Boolean.

Int(i32)

Integer.

Double(f64)

Double.

String(String)

String.

List(Vec<Value>)

List.

Map(BTreeMap<String, Value>)

Map.

Implementations

impl Value

pub fn new<V: Into<Self>>(value: V) -> Self

Creates a new Sentry value.

Panics

Panics if value is a Value::String and contains null bytes.

pub(crate) unsafe fn from_raw(raw_value: Value) -> Self

Creates a Value from sys::Value. This will deallocate the given raw_value or take ownership of it.

pub(crate) fn from_raw_borrowed(raw_value: Value) -> Self

Creates a Value from sys::Value. This will not deallocate the given raw_value.

pub(crate) fn into_raw(self) -> Value

Yields sys::Value, Value is consumed and caller is responsible for deallocating sys::Value.

Panics

Panics if self is a Value::String and contains any null bytes.

Panics if Sentry failed to allocate memory.

#[must_use]pub fn is_null(&self) -> bool

Returns true if self is Value::Null.

#[must_use]pub fn as_null(&self) -> Option<()>

Returns Some if self is Value::Null.

pub fn into_null(self) -> Result<(), Error>

Returns Ok if self is Value::Null.

Errors

Fails with Error::TryConvert if self isn't a Value::Null;

#[must_use]pub fn is_bool(&self) -> bool

Returns true if self is Value::Bool.

#[must_use]pub fn as_bool(&self) -> Option<bool>

Returns Some with the inner value if self is Value::Bool.

pub fn into_bool(self) -> Result<bool, Error>

Returns Ok with the inner value if self is Value::Bool.

Errors

Fails with Error::TryConvert if self isn't a Value::Bool;

#[must_use]pub fn is_int(&self) -> bool

Returns true if self is Value::Int.

#[must_use]pub fn as_int(&self) -> Option<i32>

Returns Some with the inner value if self is Value::Int.

pub fn into_int(self) -> Result<i32, Error>

Returns Ok with the inner value if self is Value::Int.

Errors

Fails with Error::TryConvert if self isn't a Value::Int;

#[must_use]pub fn is_double(&self) -> bool

Returns true if self is Value::Double.

#[must_use]pub fn as_double(&self) -> Option<f64>

Returns Some with the inner value if self is Value::Double.

pub fn into_double(self) -> Result<f64, Error>

Returns Ok with the inner value if self is Value::Double.

Errors

Fails with Error::TryConvert if self isn't a Value::Double;

#[must_use]pub fn is_string(&self) -> bool

Returns true if self is Value::String.

#[must_use]pub fn as_str(&self) -> Option<&str>

Returns Some with the inner value if self is Value::String.

pub fn into_string(self) -> Result<String, Error>

Returns Ok with the inner value if self is Value::String.

Errors

Fails with Error::TryConvert if self isn't a Value::String;

#[must_use]pub fn is_list(&self) -> bool

Returns true if self is Value::List.

#[must_use]pub fn as_list(&self) -> Option<&Vec<Self>>

Returns Some with the inner value if self is Value::List.

pub fn into_list(self) -> Result<Vec<Self>, Error>

Returns Ok with the inner value if self is Value::List.

Errors

Fails with Error::TryConvert if self isn't a Value::List;

#[must_use]pub fn is_map(&self) -> bool

Returns true if self is Value::Map.

#[must_use]pub fn as_map(&self) -> Option<&BTreeMap<String, Self>>

Returns Some with the inner value if self is Value::Map.

pub fn into_map(self) -> Result<BTreeMap<String, Self>, Error>

Returns Ok with the inner value if self is Value::Map.

Errors

Fails with Error::TryConvert if self isn't a Value::Map;